refactor(platform): own the name-hub records; genie consumes them from platform not daemon
Hoist `readableNameHubMethodGuards` (help / has / list / lookup /
maybeLookup) and `directoryFileMethodGuards` (makeDirectory / readText /
maybeReadText / writeText) from `@endo/daemon` into `@endo/platform/fs`,
where the rest of the portable fs vocabulary already lives. This matches
the package's purpose — decouple the filesystem types from the daemon so
any host can use them.

- platform/fs: define both records (readableNameHubMethodGuards is just
  `{ ...readableTreeMethodGuards, maybeLookup }` — the read surface is
  identical) and export them from the lite barrel.
- daemon: import both from `@endo/platform/fs/lite`; `nameHubMethodGuards`
  now spreads the imported read record and adds only the daemon-specific
  registry/locator surface.
- genie: import the records from `@endo/platform/fs/lite` instead of
  `@endo/daemon/src/interfaces.js`, and add `@endo/platform` as a direct
  dependency. genie no longer reaches into daemon internals for fs guards.

Behaviour-neutral (the records are byte-identical, just relocated);
validated by platform (255), daemon conformance + mount (88), guest/host
provisioning, and genie (15) suites, all green; platform + daemon +
genie tsc clean.
